requirements.txt
httpx>=0.27,<0.28
selectolax>=0.3.20,<0.4
lxml>=5.2,<6
python-slugify>=8.0,<9
pydantic>=2.7,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
structlog>=24.1,<25
tenacity>=9.0,<10
pypdf>=4.2,<5
pdfminer.six>=20231212
python-docx>=1.1,<2
mammoth>=1.7,<2
openpyxl>=3.1,<4
langdetect>=1.0,<2
tqdm>=4.66,<5
pytest>=8.2,<9
pytest-vcr>=1.0,<2
